+++ packages/core/src/util/styleUtils.ts
@@ -17,7 +17,12 @@ limitations under the License.
 */
 
 import Client from '../Client';
-import { DEFAULT_FONTFAMILY, DEFAULT_FONTSIZE, FONT, LINE_HEIGHT } from './Constants';
+import {
+  DEFAULT_FONTFAMILY,
+  DEFAULT_FONTSIZE,
+  FONT_STYLE_FLAG,
+  LINE_HEIGHT,
+} from './Constants';
 import Point from '../view/geometry/Point';
 import Dictionary from './Dictionary';
 import CellPath from '../view/cell/CellPath';
@@ -451,12 +456,14 @@ export const getSizeForString = (
 
   // Sets the font style
   if (fontStyle !== null) {
-    matchBinaryMask(fontStyle, FONT.BOLD) && (div.style.fontWeight = 'bold');
-    matchBinaryMask(fontStyle, FONT.ITALIC) && (div.style.fontWeight = 'italic');
+    matchBinaryMask(fontStyle, FONT_STYLE_FLAG.BOLD) && (div.style.fontWeight = 'bold');
+    matchBinaryMask(fontStyle, FONT_STYLE_FLAG.ITALIC) &&
+      (div.style.fontWeight = 'italic');
 
     const txtDecor = [];
-    matchBinaryMask(fontStyle, FONT.UNDERLINE) && txtDecor.push('underline');
-    matchBinaryMask(fontStyle, FONT.STRIKETHROUGH) && txtDecor.push('line-through');
+    matchBinaryMask(fontStyle, FONT_STYLE_FLAG.UNDERLINE) && txtDecor.push('underline');
+    matchBinaryMask(fontStyle, FONT_STYLE_FLAG.STRIKETHROUGH) &&
+      txtDecor.push('line-through');
     txtDecor.length > 0 && (div.style.textDecoration = txtDecor.join(' '));
   }